backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
geoalchemy2==0.16.0
alembic==1.14.0
pandas==2.2.3
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
openpyxl==3.1.5
pytest==8.3.4
pytest-asyncio==0.25.0